Moved all vacancy related models to model/vacancy package. Added VacancyAnalytics model
