+++ src/main/java/ru/borshchevskiy/analyticsbuilderservice/model/vacancy/SalaryEntity.java
@@ -1,4 +1,4 @@
-package ru.borshchevskiy.analyticsbuilderservice.model;
+package ru.borshchevskiy.analyticsbuilderservice.model.vacancy;
 
 import lombok.Data;
