+++ src/main/java/ru/borshchevskiy/analyticsbuilderservice/repository/VacancyRepository.java
@@ -1,7 +1,7 @@
 package ru.borshchevskiy.analyticsbuilderservice.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import ru.borshchevskiy.analyticsbuilderservice.model.VacancyEntity;
+import ru.borshchevskiy.analyticsbuilderservice.model.vacancy.VacancyEntity;
 
 import java.util.List;
